fixed register contoller bugs and setup postman

+++ src/utils/cloudinary.js
@@ -1,28 +1,39 @@
 import { v2 as cloudinary} from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv"
+dotenv.config({
+    path: './.env'
+
+});
 
 cloudinary.config(
     {
-        cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME',
-        api_key: 'process.env.CLOUDINARY_API_KEY',
-        api_secret: 'process.env.CLOUDINARY_API_SECRET'
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+        api_key: process.env.CLOUDINARY_API_KEY,
+        api_secret: process.env.CLOUDINARY_API_SECRET
     }
-)
+);
 
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!local)return null;
+        if(!localFilePath)return null;
         //upload the file on cloudinary
-        cloudinary.uploader.upload(localFilePath, {
+        const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         });
         // file has been uploaded successfully
-        console.log("file is uploaded on cloudinary", response.url);
+        fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload file failed
+            console.log("CLOUDINARY ERROR:", error);
+    
+    if (localFilePath && fs.existsSync(localFilePath)) {
+         fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload file failed
+    }
+
+    return null;
     }
 }
